requirements.txt
addict
future
lmdb
numpy>=1.21,<2.0
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
lpips
pyiqa 
natsort
chardet 
yacs 
termcolor
submitit 
tensorboardX
triton
packaging
fvcore
seaborn
pytorch_msssim
ptflops
wandb
